requirements.txt
reportlab>=4.0
Pillow>=10.0
python-docx>=1.0
